relocate `coreutils.DataWithExpiration` to `istorage/ustils.go`
