+++ pkg/istorage/utils.go
@@ -6,13 +6,16 @@
 package istorage
 
 import (
+	"encoding/binary"
 	"regexp"
 	"strconv"
 	"strings"
+	"time"
 	"unicode"
 
 	"github.com/google/uuid"
 	"github.com/voedger/voedger/pkg/appdef"
+	"github.com/voedger/voedger/pkg/coreutils/utils"
 )
 
 func NewSafeAppName(appQName appdef.AppQName, uniqueFunc func(name string) (bool, error)) (san SafeAppName, err error) {
@@ -106,4 +109,30 @@ func (san *SafeAppName) ApplyKeyspaceIsolationSuffix(suffix string) {
 		panic("too long safe name including suffix")
 	}
 	san.name += suffix
+}
+
+// first 8 bytes - ExpireAt, then - data
+func (d DataWithExpiration) ToBytes() []byte {
+	res := make([]byte, 0, len(d.Data)+utils.Uint64Size)
+	res = binary.BigEndian.AppendUint64(res, uint64(d.ExpireAt)) // nolint G115
+	res = append(res, d.Data...)
+
+	return res
+}
+
+func ReadWithExpiration(data []byte) DataWithExpiration {
+	return DataWithExpiration{
+		ExpireAt: int64(binary.BigEndian.Uint64(data[:utils.Uint64Size])), // nolint G115
+		Data:     data[utils.Uint64Size:],
+	}
+}
+
+func (d DataWithExpiration) IsExpired(now time.Time) bool {
+	return d.ExpireAt > 0 && !now.Before(time.UnixMilli(d.ExpireAt))
+}
+
+func (d DataWithExpiration) Update(data []byte) DataWithExpiration {
+	d.ExpireAt = int64(binary.BigEndian.Uint64(data[:utils.Uint64Size])) // nolint G115
+	d.Data = data[utils.Uint64Size:]
+	return d
 }
